add readme and comments in demo/gan
